requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.27
langchain-ollama==0.0.3
langgraph==0.0.26
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
icalendar==5.0.11
ollama==0.1.6
rich==13.7.1
